+++ src/main/java/io/redstudioragnarok/redcore/utils/ModReference.java
@@ -1,8 +1,11 @@
 package io.redstudioragnarok.redcore.utils;
 
 import io.redstudioragnarok.redcore.Tags;
+import io.redstudioragnarok.redcore.logging.RedLogger;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import java.net.URI;
 
 /**
  * This class defines constants for Red Core.
@@ -15,4 +18,5 @@ public class ModReference {
     public static final String NAME = "Red Core";
     public static final String VERSION = Tags.VERSION;
     public static final Logger LOG = LogManager.getLogger(ID);
+    public static final RedLogger RED_LOG = new RedLogger(NAME, "https://linkify.cz/MWCBugReport", LOG);;
 }